updated redirects and email
Updates to the login/signup process that creates a new method for managing storage of the "next" URL for signup and login processes: by storing it in session, rather than passing it as a "next" parameter.

At issue was difficulty in passing this around as a "next" parameter within querystrings. Doing this correctly was an ongoing issue, prone to breaking. For example, a user arriving at the site with parameters in the URL -- it was a challenge to encode that URL in a way that preserved those parameters, and was predictably decoded.

Storage in session stores the URL without worrying about nested encodings of parameters.

Places in which storage occurs:
  - any view using PrivateMixin
  - HTML elements that have "storeRedirect" triggered upon click

The latter is added in any location where a user action takes the user away from a page they're on, as they engage in a login or signup process – e.g. "Log in" button on a login modal, or a link that initiates a social login process.

This session "next" is then popped and used upon login or signup.

This commit also fixes an issue related to email storage, to update the User model's email field based on the email received in a social account signup.

+++ private_sharing/views.py
@@ -134,7 +134,7 @@ class JoinOnSiteDataRequestProjectView(PrivateMixin, LargePanelMixin,
 
     template_name = 'private_sharing/join-on-site.html'
 
-    def dispatch(self, *args, **kwargs):
+    def get(self, request, *args, **kwargs):
         """
         If the member has already accepted the consent form redirect them to
         the authorize page.
@@ -144,8 +144,7 @@ class JoinOnSiteDataRequestProjectView(PrivateMixin, LargePanelMixin,
                 'direct-sharing:authorize-on-site',
                 kwargs={'slug': self.get_object().slug}))
 
-        return super(JoinOnSiteDataRequestProjectView, self).dispatch(
-            *args, **kwargs)
+        return super().get(request, *args, **kwargs)
 
     # pylint: disable=unused-argument
     def post(self, request, *args, **kwargs):
@@ -206,7 +205,7 @@ class AuthorizeOnSiteDataRequestProjectView(PrivateMixin, LargePanelMixin,
 
     template_name = 'private_sharing/authorize-on-site.html'
 
-    def dispatch(self, *args, **kwargs):
+    def get(self, request, *args, **kwargs):
         """
         If the member hasn't already accepted the consent form redirect them to
         the consent form page.
@@ -217,11 +216,17 @@ class AuthorizeOnSiteDataRequestProjectView(PrivateMixin, LargePanelMixin,
                 'direct-sharing:join-on-site',
                 kwargs={'slug': self.get_object().slug}))
 
-        return super().dispatch(
-            *args, **kwargs)
+        return super().get(request, *args, **kwargs)
 
     # pylint: disable=unused-argument
     def post(self, request, *args, **kwargs):
+        # repeating this because making a function for these two lines
+        # would add more complexity than it would save.
+        if not self.project_joined_by_member:
+            return HttpResponseRedirect(reverse_lazy(
+                'direct-sharing:join-on-site',
+                kwargs={'slug': self.get_object().slug}))
+
         if self.request.POST.get('cancel') == 'cancel':
             self.project_member.delete()
